--- helpers/element_creaters/createNewElement.js
@@ -1,7 +1,0 @@
-// export function createNewElement(tag, className) {
-//     const $tag = document.createElement(tag)
-//     if (className) {
-//         $tag.classList.add(className)
-//     }
-//     return $tag
-// }
