--- helpers/getRandomNumber.js
@@ -1,1 +1,0 @@
-// export const getRandomNumber = (num) => Math.ceil(Math.random() * num)
