--- helpers/helth_points/changeHP.js
@@ -1,8 +1,0 @@
-// export function changeHP(hitPoints) {
-//     this.hp -= hitPoints
-//
-//     if (this.hp <= 0) {
-//         return this.hp = 0
-//     }
-//     return this.hp
-// }
