--- helpers/helth_points/elementHP.js
@@ -1,3 +1,0 @@
-// export function elementHP() {
-//     return document.querySelector(`.player${this.player} .life`)
-// }
