--- helpers/helth_points/renderHP.js
@@ -1,3 +1,0 @@
-// export function renderHP() {
-//     this.elementHP().style.width = this.hp + '%'
-// }
