--- helpers/logs_and_banners/commentInChat.js
@@ -1,1 +1,0 @@
-// export const commentInChat = (time,text, damage, hp) => `<p>${time} ${text} ${damage} ${hp}</p>`
