+++ main.js
@@ -1,13 +1,3 @@
-import {player1, player2} from "./store/players.js";
-import {createNewElement} from "./helpers/element_creaters/createNewElement.js";
-import {playerAttack} from "./helpers/action_functions/playerAttack.js";
-import {enemyAttack} from "./helpers/action_functions/enemyAttck.js";
-import {changeHPlayers} from "./helpers/helth_points/changeHPPlayers.js";
-import {startBattle} from "./helpers/logs_and_banners/startBanner.js";
-import {showResult} from "./helpers/logs_and_banners/showResult.js";
-import Game from "./GAME.js";
-
-
 // export const $arenas = document.querySelector('.arenas')
 // export const $chat = document.querySelector('.chat')
 // export const $randomButton = document.querySelector('.button')
@@ -16,40 +6,9 @@ import Game from "./GAME.js";
 // $arenas.appendChild(createPlayer(player1))
 // $arenas.appendChild(createPlayer(player2))
 
-
-$controlForm.addEventListener('submit', function (event) {
-    event.preventDefault();
-
-    const playerHitPoints = playerAttack()
-    const enemyHitPoint = enemyAttack()
-
-    changeHPlayers(playerHitPoints, enemyHitPoint)
-    showResult()
-
-})
-
-// export function createPlayer(playerC) {
-//     const player1 = createNewElement("div", "player" + playerC.player)
-//     const progressbar = createNewElement("div", "progressbar")
-//     const life = createNewElement("div", "life")
-//     const name = createNewElement("div", "name")
-//     const character = createNewElement("div", "character")
-//     const img = createNewElement('img', "img")
-//
-//     progressbar.appendChild(name)
-//     progressbar.appendChild(life)
-//     character.appendChild(img)
-//     player1.appendChild(progressbar)
-//     player1.appendChild(character)
-//
-//     life.style.width = playerC.hp + "%"
-//     name.innerText = playerC.name
-//     img.src = playerC.img
-//
-//     return player1
-// }
-
-startBattle(player1, player2)
+import Game from "./classes/GAME.js";
 
 const game = new Game();
+// game.startBattle(player1, player2)
+
 game.start()
